auth: encapsulate web redirect bootstrap inside auth driven-impl
Move callback URL detection, OIDC response persistence, and history
replacement into a `WebAuthBootstrap` object in
`feat/authentication/fe/driven/impl/webMain`. composeApp's main.kt
becomes a 2-line dispatch that no longer imports oidc-preferences,
ktor URL types, or `WebRunConfig`.
